Feature authorization delete product in admin

+++ controllers/controllerUser.js
@@ -1,11 +1,12 @@
-const { User, Product } = require("../models");
+const { User, Product, Category} = require("../models");
 const { signToken } = require("../helpers/jwt");
 const { comparePassword } = require("../helpers/bCrypt");
+const sendEmail = require('../helpers/nodemailer')
 
 class ControllerUser {
   static async register(req, res, next) {
     try {
-      const { name, username, email, password, address, phoneNumber, gender } =
+      const { name, username, email, password, address, phoneNumber, gender} =
         req.body;
       const payload = {
         name,
@@ -15,11 +16,13 @@ class ControllerUser {
         address,
         phoneNumber,
         gender,
+        role: 'Admin'
       };
       const result = await User.create(payload);
-      res
-        .status(201)
-        .json({ id: result.id, name: result.name, email: result.email });
+      const content = `Hi ${payload.name}!, your account with email ${payload.email} successfully registered.`
+      const subject = `Information Registered`
+      sendEmail(result, content, subject)
+      res.status(201).json({ id: result.id, name: result.name, email: result.email });
     } catch (err) {
       next(err);
     }
@@ -47,12 +50,14 @@ class ControllerUser {
   static async addProduct(req, res, next) {
     try {
       const { name, price, weight, imageUrl, categoryId } = req.body;
+      const {id} = req.user
       const payload = {
         name,
         price: +price,
         weight: +weight,
         imageUrl,
         categoryId: +categoryId,
+        authorId: +id
       };
       const newProduct = await Product.create(payload)
       res.status(201).json(newProduct);
@@ -62,7 +67,20 @@ class ControllerUser {
   }
   static async getAllProducts(req, res, next) {
     try {
-      const result = await Product.findAll();
+      const result = await Product.findAll(
+        {
+        attributes: {exclude: ['createdAt', 'updatedAt', ]},
+        include: [{
+          model: Category,
+          attributes: {exclude: ['id', 'createdAt', 'updatedAt']},
+        },
+        {
+          model: User,
+          attributes: {exclude: ['createdAt', 'updatedAt', 'id', 'username', 'password', 'address', 'gender', 'role']},
+        }
+      ]
+      }
+      );
       res.status(200).json(result);
     } catch (err) {
       next(err)
@@ -71,11 +89,18 @@ class ControllerUser {
   static async deleteProduct(req, res, next) {
     try {
       const productId = Number(req.params.productId)
+      if(!productId) {
+        throw {name: 'bad request'}
+      }
+      const product = await Product.findOne({ where: { id: productId || null } });
+      if (!product) {
+        throw {name: 'NotFound'}
+      }
       const deleteProduct = await Product.destroy({ where: { id: productId } });
       if(!deleteProduct) {
         throw {name: 'NotFound'}
       }
-      res.status(200).json({message: `product  with name ${deleteProduct.name} & id ${deleteProduct.id} has been deleted`})
+      res.status(200).json({message: `Product has been Deleted`})
     } catch (err) {
       next(err)
     }
